# docs/swagger.yaml
definitions:
  model.Baggage:
    properties:
      airline:
        example: AirlineX
        type: string
      baggage_code:
        example: ABC123
        type: string
      baggage_id:
        type: integer
      baggage_status:
        example: checked
        type: string
      baggage_type:
        example: suitcase
        type: string
      owner_name:
        example: John Doe
        type: string
      pasport_details:
        example: "123456789"
        type: string
      photo:
        example: http://example.com/baggage.jpg
        type: string
      size:
        example: large
        type: string
      weight:
        example: 23.5
        type: number
    type: object
  model.BaggageRequest:
    properties:
      airline:
        example: AirlineX
        type: string
      baggage_code:
        example: ABC123
        type: string
      baggage_type:
        example: suitcase
        type: string
      owner_name:
        example: John Doe
        type: string
      pasport_details:
        example: "123456789"
        type: string
      size:
        example: large
        type: string
      weight:
        example: 23.5
        type: number
    type: object
  model.BaggagesGetResponse:
    properties:
      baggages:
        items:
          $ref: '#/definitions/model.Baggage'
        type: array
      delivery_id:
        example: 1
        type: integer
    type: object
...
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /baggage:
    get:
      description: Возращает список всех активных багажей
      parameters:
      - description: Код багажа
        format: email
        in: query
        name: searchCode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список багажей
          schema:
            $ref: '#/definitions/model.BaggagesGetResponse'
        "400":
          description: Обработанная ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Получение списка багажа
      tags:
      - Багаж
  /baggage/{baggage_id}:
    get:
      description: Возвращает информацию о багаже по его ID
      parameters:
      - description: ID багажа
        in: path
        name: baggage_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о багаже
          schema:
            $ref: '#/definitions/model.Baggage'
        "400":
          description: Обработанная ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Получение багажа по ID
      tags:
      - Багаж
  /baggage/{baggage_id}/delete:
    delete:
      description: Удаляет багаж по его ID
      parameters:
      - description: ID багажа
        in: path
        name: baggage_id
        required: true
        type: integer
      - description: Код багажа
        format: email
        in: query
        name: searchCode
        type: string
      - description: Токен авторизации
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список багажей
          schema:
            $ref: '#/definitions/model.BaggagesGetResponse'
        "400":
          description: Обработанная ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: У пользователя нет прав для этого запроса
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Удаление багажа
      tags:
      - Багаж
...
  /baggage/{baggage_id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Добавляет изображение к багажу по его ID
      parameters:
      - description: ID багажа
        in: path
        name: baggage_id
        required: true
        type: integer
      - description: Изображение багажа
        in: formData
        name: image
        required: true
        type: file
      - description: Токен авторизации
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о багаже с изображением
          schema:
            $ref: '#/definitions/model.Baggage'
        "400":
          description: Обработанная ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: У пользователя нет прав для этого запроса
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Добавление изображения к багажу
      tags:
      - Багаж
  /baggage/{baggage_id}/update:
    put:
      consumes:
      - application/json
      description: Обновляет информацию о багаже по его ID
      parameters:
      - description: ID багажа
        in: path
        name: baggage_id
        required: true
        type: integer
      - description: Токен авторизации
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о багаже
          schema:
            $ref: '#/definitions/model.Baggage'
        "400":
          description: Обработанная ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: У пользователя нет прав для этого запроса
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Обновление информации о багаже
      tags:
      - Багаж
  /baggage/create:
    post:
      consumes:
      - application/json
      description: Создает новый багаж с предоставленными данными
      parameters:
      - description: Код багажа
        format: email
        in: query
        name: searchCode
        type: string
      - description: Пользовательский объект в формате JSON
        in: body
        name: baggage
        required: true
        schema:
          $ref: '#/definitions/model.BaggageRequest'
      - description: Токен авторизации
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список багажей
          schema:
            $ref: '#/definitions/model.BaggagesGetResponse'
        "400":
          description: Обработанная ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: У пользователя нет прав для этого запроса
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Создание нового багажа
      tags:
      - Багаж
...
swagger: "2.0"